modify boot script

+++ script/clientboot.sh
@@ -1,23 +1,32 @@
 #!/bin/bash
 
-touch /home/linaro/clientboot.txt
-
-sleep 5
+echo "kill wpa_supplicant"
 sudo killall wpa_supplicant
-sleep 1
+sleep 0.5
+echo "connect wifi ap"
 sudo /sbin/wpa_supplicant -B -i wlan0 -c /etc/wpa_supplicant/wpa_supplicant.conf
 
 if [[ -e '/home/linaro/apboot.txt' ]]; then
-	sleep 10
 	echo "check network"
-	sudo ping www.google.com -w 5
-	if [[ $? == 0 ]]; then
-		echo "network ok, remove apboot.txt"
-		rm /home/linaro/apboot.txt
-	else
-		echo "network fail, return AP mode"
-		/home/linaro/init/setmode.sh ap
-		sync
-		sudo reboot
-	fi
+	for var in $(seq 1 6)
+	do
+		sleep 10
+		sudo ping www.google.com -w 5
+		if [[ $? == 0 ]]; then
+			echo "network ok, remove apboot.txt, make clientboot.txt"
+			rm /home/linaro/apboot.txt
+			touch /home/linaro/clientboot.txt
+			exit 0
+		else
+			echo "var = $var"
+			if [[ $var = 6 ]]; then
+				echo "network fail, return AP mode"
+				/home/linaro/init/setmode.sh ap
+				sync
+				sudo reboot
+			fi
+		fi
+	done
+else
+	touch /home/linaro/clientboot.txt
 fi
